+++ src/test/samples/TestLibrary.groovy
@@ -1,0 +1,12 @@
+library(
+  namespace: 'hubitat-vscode-ext',
+  author: 'Unknown',
+  category: 'samples',
+  description: 'Some description',
+  name: 'Sample Library',
+  documentationLink: ''
+)
+
+void testStub1() {
+  log.debug('I am a sample library')';'
+}