+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 All notable changes to the "hubitat-dev" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
 ## [Unreleased]
 
+## [0.0.10]
+Hides Hubitat status bar for non .groovy files
+Fixed progress spinner on updating code
+Logged better error messages when failing to add a NEW code(driver,app,library) element
+##
 - Initial release